BJS2-21415 bug fixes

+++ src/main/java/faang/school/godbless/BJS2_21415/Calculator.java
@@ -2,5 +2,5 @@ package faang.school.godbless.BJS2_21415;
 
 @FunctionalInterface
 public interface Calculator {
-    int calc(int a, int b);
+    long calc(long a, int b);
 }
